build fixen

+++ src/ont/user-2d.c
@@ -694,8 +694,8 @@ static void draw_links(char* path, uint8_t container_g2d_node){
 
     if(i==swipe_index){
       uint16_t droppables=object_property_length(user, "inventory:list");
-      uint8_t drop_col = droppables? G2D_BLUE: G2D_GREY_7;
-      char*   drop_text   = droppables? "O->": "---";
+      uint16_t drop_col  = droppables? G2D_BLUE: G2D_GREY_7;
+      char*    drop_text = droppables? "O->": "---";
       draw_swipe_feedback(container_g2d_node, y, G2D_RED, G2D_GREEN_18, drop_col,
                                                  "<-X",   "<-O",        drop_text);
     }
@@ -780,8 +780,8 @@ static uint8_t make_in_scroll_button(uint8_t scroll_g2d_node,
 
   if(control==swipe_control){
     uint16_t droppables=object_property_length(user, "inventory:list");
-    uint8_t drop_col = droppables? G2D_BLUE: G2D_GREY_7;
-    char*   drop_text   = droppables? "O->": "---";
+    uint16_t drop_col  = droppables? G2D_BLUE: G2D_GREY_7;
+    char*    drop_text = droppables? "O->": "---";
     draw_swipe_feedback(scroll_g2d_node, y, G2D_GREEN_18, G2D_GREEN_18, drop_col,
                                             "<-O",        "<-O",        drop_text);
   }
@@ -883,8 +883,8 @@ static void draw_list(char* path, uint8_t g2d_node) {
 
     if(i==swipe_index){
       uint16_t droppables=object_property_length(user, "inventory:list");
-      uint8_t drop_col = droppables? G2D_BLUE: G2D_GREY_7;
-      char*   drop_text   = droppables? "O->": "---";
+      uint16_t drop_col  = droppables? G2D_BLUE: G2D_GREY_7;
+      char*    drop_text = droppables? "O->": "---";
       draw_swipe_feedback(scroll_g2d_node, y, G2D_RED, G2D_GREEN_18, drop_col,
                                               "<-X",   "<-O",        drop_text);
     }
